pyproject.toml
[project]
name = "favicon-thief"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pillow>=11.1.0",
    "requests>=2.32.3",
]
